feat(hydration): move useHydrate and Hydrate into react package
