+++ src/hydration/index.ts
@@ -1,5 +1,8 @@
-export { dehydrate, hydrate } from 'react-query'
-export { useHydrate, Hydrate } from './react'
+// This package once contained these functions, but they have now been moved
+// into the core and react packages.
+// They are re-exported here to avoid a breaking change, but this package
+// should be considered deprecated and removed in a future major version.
+export { dehydrate, hydrate, useHydrate, Hydrate } from 'react-query'
 
 // Types
 export type {
@@ -9,4 +12,4 @@ export type {
   ShouldDehydrateMutationFunction,
   ShouldDehydrateQueryFunction,
 } from '../core/hydration'
-export type { HydrateProps } from './react'
+export type { HydrateProps } from '../react/Hydrate'
