+++ src/react/index.ts
@@ -13,8 +13,10 @@ export { useMutation } from './useMutation'
 export { useQuery } from './useQuery'
 export { useQueries } from './useQueries'
 export { useInfiniteQuery } from './useInfiniteQuery'
+export { useHydrate, Hydrate } from './Hydrate'
 
 // Types
 export * from './types'
 export type { QueryClientProviderProps } from './QueryClientProvider'
 export type { QueryErrorResetBoundaryProps } from './QueryErrorResetBoundary'
+export type { HydrateProps } from './Hydrate'
